Restructure api handlers --closes #12
