+++ controllers/auth.js
@@ -1,5 +1,5 @@
 import express from 'express'
-import * as apiHandler from '../handlers/api'
+import * as slackApi from '../handlers/api/slackApi'
 import * as storeHandler from '../handlers/store'
 import * as botHandler from '../handlers/bot'
 
@@ -7,7 +7,7 @@ const router = new express.Router()
 
 router.get('/auth', async function (req, res) {
   try {
-    let token = await apiHandler.exchangeCodeForToken(req.query.code)
+    let token = await slackApi.exchangeCodeForToken(req.query.code)
     storeHandler.storeTeamToken(token)
     botHandler.createNewBotConnection({
       token: token.bot.bot_access_token,
