+++ handlers/api/pathsApi.js
