+++ handlers/store.js
@@ -1,5 +1,5 @@
 import firebase from 'firebase'
-import * as apiHandler from './api'
+import * as slackApi from './api/slackApi'
 
 const TOKENS = 'tokens'
 
@@ -29,7 +29,7 @@ const storeTeamToken = (token) => {
 }
 
 const setupDevTeam = async function() {
-  let devBotData = await apiHandler.identifyDevBotData()
+  let devBotData = await slackApi.identifyDevBotData()
   const botData = { bot_access_token: process.env.slack_bot_token, bot_user_id: devBotData.user_id}
   const tokenData = { bot: botData, team_id: devBotData.team_id, access_token: process.env.slack_api_token}
   storeTeamToken(tokenData)
